pyproject.toml
[project]
name = "brief-spark-structured-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "confluent-kafka>=2.12.2",
    "delta-spark>=4.0.0",
]
